+++ src/main/java/com/example/sbkafka/Repository/OrderFormRepository.java
@@ -19,28 +19,28 @@ public interface OrderFormRepository extends JpaRepository<OrderForm,Integer> {
 	public int countOfOrders();
 	
 	@Query("select sum(calc) from OrderForm f")//сумма всех заказов без ндс
-	public double sumOfOrders();
+	public Double sumOfOrders();
 	
 	@Query("select min(calc) from OrderForm f")//минимальный заказ
-	public double minOfOrder();
+	public Double minOfOrder();
 	
 	@Query("select max(calc) from OrderForm f")//максимальный заказ
-	public double maxOfOrder();
+	public Double maxOfOrder();
 	
 	@Query("select avg(calc) from OrderForm f")//средняя стоимость заказов
-	public double avgOfOrder();
+	public Double avgOfOrder();
 	
 	@Query("select count(f) from OrderForm f where calc in(:f1,:f2,:f3,:f4,:f5,:f6)")//кол-во ложных выездов до КАД
 	public int countOfFalseOut(@Param("f1") double f1,@Param("f2") double f2,@Param("f3") double f3,@Param("f4") double f4,@Param("f5") double f5,@Param("f6") double f6);
 	
 	@Query("select sum(calc) from OrderForm f where calc in(:f1,:f2,:f3,:f4,:f5,:f6)")//сумма ложных выездов до КАД
-	public int sumOfFalseOut(@Param("f1") double f1,@Param("f2") double f2,@Param("f3") double f3,@Param("f4") double f4,@Param("f5") double f5,@Param("f6") double f6);
+	public Double sumOfFalseOut(@Param("f1") double f1,@Param("f2") double f2,@Param("f3") double f3,@Param("f4") double f4,@Param("f5") double f5,@Param("f6") double f6);
 	
 	@Query("select count(f) from OrderForm f where comm in('Fix')")//кол-во заказов ДФС
 	public int countDfs();
 	
 	@Query("select sum(calc) from OrderForm f where comm in('Fix')")//сумма заказов ДФС
-	public int sumDfs();
+	public Double sumDfs();
 	
 	@Query(value="select * from orderform where ordernumber in(:ordernum) order by ordernumber",nativeQuery=true)//поиск по номеру заказа
 	public List<OrderForm> findByOrderNumber(@Param("ordernum") int ordernum);
